Revert "refactor log service (#292)" (#293)
This reverts commit e44d465b30e2b14c61f941929c6c3ce5b8c1c252.

+++ logservice/logpuller/log_puller.go
@@ -181,7 +181,7 @@ func (p *LogPuller) Subscribe(
 	p.subscriptions.spanProgressMap[subID] = progress
 	p.subscriptions.Unlock()
 
-	p.client.Subscribe(subID, span, startTs)
+	p.client.Subscribe(subID, span, uint64(startTs))
 	return subID
 }
